+++ HRSTransport/GDHTransLargeX0/GDHTransLargeX0.cc
@@ -79,7 +79,7 @@ GDHTransLargeX0::~GDHTransLargeX0()
 /////////////////////////////////////////////////////////////////////////////
 bool GDHTransLargeX0::TransRightHRS(double* pV5)
 {
-	float vector_jjl[]={pV5[0],pV5[1],pV5[2],pV5[3],pV5[4]};
+	float vector_jjl[]={(float)pV5[0],(float)pV5[1],(float)pV5[2],(float)pV5[3],(float)pV5[4]};
 	float x_test, y_test;
 	int ii=5;
 
@@ -173,7 +173,7 @@ bool GDHTransLargeX0::TransRightHRS(double* pV5)
 bool GDHTransLargeX0::TransLeftHRS(double* pV5)
 {
 	//in order to call right arm routines, need to flip y, phi 
-	float vector_jjl[]={pV5[0],pV5[1],-pV5[2],-pV5[3],pV5[4]};
+	float vector_jjl[]={(float)pV5[0],(float)pV5[1],-(float)pV5[2],-(float)pV5[3],(float)pV5[4]};
 	bool bGoodParticle=TransRightHRS(pV5);
 	if(bGoodParticle)
 	{
@@ -207,7 +207,7 @@ bool GDHTransLargeX0::TransLeftHRS(double* pV5)
 /////////////////////////////////////////////////////////////////////////////////////
 void GDHTransLargeX0::ReconRightHRS(double *pV5)
 {
-	float vector_jjl[]={pV5[0],pV5[1],pV5[2],pV5[3],pV5[4]};
+	float vector_jjl[]={(float)pV5[0],(float)pV5[1],(float)pV5[2],(float)pV5[3],(float)pV5[4]};
 	int ii=5,jj=1;
 
 	/* Orthogonalize theta as JJL asks*/
@@ -233,7 +233,7 @@ void GDHTransLargeX0::ReconLeftHRS(double *pV5)
 	//routines for the 2nd version of GDH experiment
 	
 	//in order to call right arm routines, need to flip y, phi 
-	float vector_jjl[]={pV5[0],pV5[1],-pV5[2],-pV5[3],pV5[4]};
+	float vector_jjl[]={(float)pV5[0],(float)pV5[1],-(float)pV5[2],-(float)pV5[3],(float)pV5[4]};
 	int ii=5,jj=1;
 
 	/* Orthogonalize theta as JJL asks*/
